% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{abort_environment_update}
\alias{abort_environment_update}
\title{Cancels in-progress environment configuration update or application version deployment}
\usage{
abort_environment_update(EnvironmentId = NULL, EnvironmentName = NULL)
}
\arguments{
\item{EnvironmentId}{This specifies the ID of the environment with the in-progress update that you want to cancel.}

\item{EnvironmentName}{This specifies the name of the environment with the in-progress update that you want to cancel.}
}
\description{
Cancels in-progress environment configuration update or application version deployment.
}
\examples{
# The following code aborts a running application version deployment for
# an environment named my-env:
abort_environment_update(
  EnvironmentName = "my-env"
)

}
