% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{apply_environment_managed_action}
\alias{apply_environment_managed_action}
\title{Applies a scheduled managed action immediately}
\usage{
apply_environment_managed_action(ActionId, EnvironmentName = NULL,
  EnvironmentId = NULL)
}
\arguments{
\item{ActionId}{The action ID of the scheduled managed action to execute.}

\item{EnvironmentName}{The name of the target environment.}

\item{EnvironmentId}{The environment ID of the target environment.}
}
\description{
Applies a scheduled managed action immediately. A managed action can be applied only if its status is \code{Scheduled}. Get the status and action ID of a managed action with DescribeEnvironmentManagedActions.
}
\examples{

}
