% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{delete_application}
\alias{delete_application}
\title{Deletes the specified application along with all associated versions and configurations}
\usage{
delete_application(ApplicationName, TerminateEnvByForce = NULL)
}
\arguments{
\item{ApplicationName}{[required] The name of the application to delete.}

\item{TerminateEnvByForce}{When set to true, running environments will be terminated before deleting the application.}
}
\description{
Deletes the specified application along with all associated versions and configurations. The application versions will not be deleted from your Amazon S3 bucket.
}
\details{
You cannot delete an application that has a running environment.
}
\section{Accepted Parameters}{
\preformatted{delete_application(
  ApplicationName = "string",
  TerminateEnvByForce = TRUE|FALSE
)
}
}

\examples{
# The following operation deletes an application named my-app:
\donttest{delete_application(
  ApplicationName = "my-app"
)}

}
