% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{delete_configuration_template}
\alias{delete_configuration_template}
\title{Deletes the specified configuration template}
\usage{
delete_configuration_template(ApplicationName, TemplateName)
}
\arguments{
\item{ApplicationName}{[required] The name of the application to delete the configuration template from.}

\item{TemplateName}{[required] The name of the configuration template to delete.}
}
\description{
Deletes the specified configuration template.
}
\details{
When you launch an environment using a configuration template, the environment gets a copy of the template. You can delete or modify the environment's copy of the template without affecting the running environment.
}
\section{Accepted Parameters}{
\preformatted{delete_configuration_template(
  ApplicationName = "string",
  TemplateName = "string"
)
}
}

\examples{
# The following operation deletes a configuration template named
# my-template for an application named my-app:
\donttest{delete_configuration_template(
  ApplicationName = "my-app",
  TemplateName = "my-template"
)}

}
