% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{delete_platform_version}
\alias{delete_platform_version}
\title{Deletes the specified version of a custom platform}
\usage{
delete_platform_version(PlatformArn = NULL)
}
\arguments{
\item{PlatformArn}{The ARN of the version of the custom platform.}
}
\description{
Deletes the specified version of a custom platform.
}
\examples{

}
