% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{describe_account_attributes}
\alias{describe_account_attributes}
\title{Returns attributes related to AWS Elastic Beanstalk that are associated with the calling AWS account}
\usage{
describe_account_attributes()
}
\description{
Returns attributes related to AWS Elastic Beanstalk that are associated with the calling AWS account.
}
\details{
The result currently has one set of attributes---resource quotas.
}
