% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{describe_applications}
\alias{describe_applications}
\title{Returns the descriptions of existing applications}
\usage{
describe_applications(ApplicationNames = NULL)
}
\arguments{
\item{ApplicationNames}{If specified, AWS Elastic Beanstalk restricts the returned descriptions to only include those with the specified names.}
}
\description{
Returns the descriptions of existing applications.
}
\section{Accepted Parameters}{
\preformatted{describe_applications(
  ApplicationNames = list(
    "string"
  )
)
}
}

\examples{
# The following operation retrieves information about applications in the
# current region:
\donttest{describe_applications()}

}
