% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{describe_environment_health}
\alias{describe_environment_health}
\title{Returns information about the overall health of the specified environment}
\usage{
describe_environment_health(EnvironmentName = NULL, EnvironmentId = NULL,
  AttributeNames = NULL)
}
\arguments{
\item{EnvironmentName}{Specify the environment by name.

You must specify either this or an EnvironmentName, or both.}

\item{EnvironmentId}{Specify the environment by ID.

You must specify either this or an EnvironmentName, or both.}

\item{AttributeNames}{Specify the response elements to return. To retrieve all attributes, set to \code{All}. If no attribute names are specified, returns the name of the environment.}
}
\description{
Returns information about the overall health of the specified environment. The \strong{DescribeEnvironmentHealth} operation is only available with AWS Elastic Beanstalk Enhanced Health.
}
\examples{
# The following operation retrieves overall health information for an
# environment named my-env:
describe_environment_health(
  AttributeNames = list(
    "All"
  ),
  EnvironmentName = "my-env"
)

}
