% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{describe_environment_managed_action_history}
\alias{describe_environment_managed_action_history}
\title{Lists an environment's completed and failed managed actions}
\usage{
describe_environment_managed_action_history(EnvironmentId = NULL,
  EnvironmentName = NULL, NextToken = NULL, MaxItems = NULL)
}
\arguments{
\item{EnvironmentId}{The environment ID of the target environment.}

\item{EnvironmentName}{The name of the target environment.}

\item{NextToken}{The pagination token returned by a previous request.}

\item{MaxItems}{The maximum number of items to return for a single request.}
}
\description{
Lists an environment's completed and failed managed actions.
}
\examples{

}
