% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{describe_environment_managed_actions}
\alias{describe_environment_managed_actions}
\title{Lists an environment's upcoming and in-progress managed actions}
\usage{
describe_environment_managed_actions(EnvironmentName = NULL,
  EnvironmentId = NULL, Status = NULL)
}
\arguments{
\item{EnvironmentName}{The name of the target environment.}

\item{EnvironmentId}{The environment ID of the target environment.}

\item{Status}{To show only actions with a particular status, specify a status.}
}
\description{
Lists an environment's upcoming and in-progress managed actions.
}
\examples{

}
