% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{describe_environment_resources}
\alias{describe_environment_resources}
\title{Returns AWS resources for this environment}
\usage{
describe_environment_resources(
  EnvironmentId = "string",
  EnvironmentName = "string"
)
}
\arguments{
\item{EnvironmentId}{The ID of the environment to retrieve AWS resource usage data.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns \code{MissingRequiredParameter} error.}

\item{EnvironmentName}{The name of the environment to retrieve AWS resource usage data.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns \code{MissingRequiredParameter} error.}
}
\description{
Returns AWS resources for this environment.
}
\examples{
# The following operation retrieves information about resources in an
# environment named my-env:
describe_environment_resources(
  EnvironmentName = "my-env"
)

}
