% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{describe_platform_version}
\alias{describe_platform_version}
\title{Describes the version of the platform}
\usage{
describe_platform_version(
  PlatformArn = "string"
)
}
\arguments{
\item{PlatformArn}{The ARN of the version of the platform.}
}
\description{
Describes the version of the platform.
}
\examples{

}
