% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{list_platform_versions}
\alias{list_platform_versions}
\title{Lists the available platforms}
\usage{
list_platform_versions(Filters = NULL, MaxRecords = NULL,
  NextToken = NULL)
}
\arguments{
\item{Filters}{List only the platforms where the platform member value relates to one of the supplied values.}

\item{MaxRecords}{The maximum number of platform values returned in one call.}

\item{NextToken}{The starting index into the remaining list of platforms. Use the \code{NextToken} value from a previous \code{ListPlatformVersion} call.}
}
\description{
Lists the available platforms.
}
\section{Accepted Parameters}{
\preformatted{list_platform_versions(
  Filters = list(
    list(
      Type = "string",
      Operator = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  NextToken = "string"
)
}
}

