% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_package.R
\docType{package}
\name{paws.elasticbeanstalk-package}
\alias{paws.elasticbeanstalk}
\alias{paws.elasticbeanstalk-package}
\title{paws.elasticbeanstalk: AWS Elastic Beanstalk}
\description{
AWS Elastic Beanstalk makes it easy for you to
    create, deploy, and manage scalable, fault-tolerant applications
    running on the Amazon Web Services cloud.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
