% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{restart_app_server}
\alias{restart_app_server}
\title{Causes the environment to restart the application container server running on each Amazon EC2 instance}
\usage{
restart_app_server(EnvironmentId = NULL, EnvironmentName = NULL)
}
\arguments{
\item{EnvironmentId}{The ID of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns \code{MissingRequiredParameter} error.}

\item{EnvironmentName}{The name of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns \code{MissingRequiredParameter} error.}
}
\description{
Causes the environment to restart the application container server running on each Amazon EC2 instance.
}
\section{Accepted Parameters}{
\preformatted{restart_app_server(
  EnvironmentId = "string",
  EnvironmentName = "string"
)
}
}

\examples{
# The following operation restarts application servers on all instances in
# an environment named my-env:
\donttest{restart_app_server(
  EnvironmentName = "my-env"
)}

}
