% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{swap_environment_cnam_es}
\alias{swap_environment_cnam_es}
\title{Swaps the CNAMEs of two environments}
\usage{
swap_environment_cnam_es(SourceEnvironmentId = NULL,
  SourceEnvironmentName = NULL, DestinationEnvironmentId = NULL,
  DestinationEnvironmentName = NULL)
}
\arguments{
\item{SourceEnvironmentId}{The ID of the source environment.

Condition: You must specify at least the \code{SourceEnvironmentID} or the \code{SourceEnvironmentName}. You may also specify both. If you specify the \code{SourceEnvironmentId}, you must specify the \code{DestinationEnvironmentId}.}

\item{SourceEnvironmentName}{The name of the source environment.

Condition: You must specify at least the \code{SourceEnvironmentID} or the \code{SourceEnvironmentName}. You may also specify both. If you specify the \code{SourceEnvironmentName}, you must specify the \code{DestinationEnvironmentName}.}

\item{DestinationEnvironmentId}{The ID of the destination environment.

Condition: You must specify at least the \code{DestinationEnvironmentID} or the \code{DestinationEnvironmentName}. You may also specify both. You must specify the \code{SourceEnvironmentId} with the \code{DestinationEnvironmentId}.}

\item{DestinationEnvironmentName}{The name of the destination environment.

Condition: You must specify at least the \code{DestinationEnvironmentID} or the \code{DestinationEnvironmentName}. You may also specify both. You must specify the \code{SourceEnvironmentName} with the \code{DestinationEnvironmentName}.}
}
\description{
Swaps the CNAMEs of two environments.
}
\examples{
# The following operation swaps the assigned subdomains of two
# environments:
swap_environment_cnam_es(
  DestinationEnvironmentName = "my-env-green",
  SourceEnvironmentName = "my-env-blue"
)

}
