% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{update_application}
\alias{update_application}
\title{Updates the specified application to have the specified properties}
\usage{
update_application(ApplicationName, Description = NULL)
}
\arguments{
\item{ApplicationName}{The name of the application to update. If no such application is found, \code{UpdateApplication} returns an \code{InvalidParameterValue} error.}

\item{Description}{A new description for the application.

Default: If not specified, AWS Elastic Beanstalk does not update the description.}
}
\description{
Updates the specified application to have the specified properties.
}
\details{
If a property (for example, \code{description}) is not provided, the value remains unchanged. To clear these properties, specify an empty string.
}
\examples{
# The following operation updates the description of an application named
# my-app:
update_application(
  ApplicationName = "my-app",
  Description = "my Elastic Beanstalk application"
)

}
