% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{update_application_resource_lifecycle}
\alias{update_application_resource_lifecycle}
\title{Modifies lifecycle settings for an application}
\usage{
update_application_resource_lifecycle(ApplicationName,
  ResourceLifecycleConfig)
}
\arguments{
\item{ApplicationName}{[required] The name of the application.}

\item{ResourceLifecycleConfig}{[required] The lifecycle configuration.}
}
\description{
Modifies lifecycle settings for an application.
}
\section{Accepted Parameters}{
\preformatted{update_application_resource_lifecycle(
  ApplicationName = "string",
  ResourceLifecycleConfig = list(
    ServiceRole = "string",
    VersionLifecycleConfig = list(
      MaxCountRule = list(
        Enabled = TRUE|FALSE,
        MaxCount = 123,
        DeleteSourceFromS3 = TRUE|FALSE
      ),
      MaxAgeRule = list(
        Enabled = TRUE|FALSE,
        MaxAgeInDays = 123,
        DeleteSourceFromS3 = TRUE|FALSE
      )
    )
  )
)
}
}

