% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{validate_configuration_settings}
\alias{validate_configuration_settings}
\title{Takes a set of configuration settings and either a configuration template or environment, and determines whether those values are valid}
\usage{
validate_configuration_settings(ApplicationName, OptionSettings,
  TemplateName = NULL, EnvironmentName = NULL)
}
\arguments{
\item{ApplicationName}{The name of the application that the configuration template or environment belongs to.}

\item{OptionSettings}{A list of the options and desired values to evaluate.}

\item{TemplateName}{The name of the configuration template to validate the settings against.

Condition: You cannot specify both this and an environment name.}

\item{EnvironmentName}{The name of the environment to validate the settings against.

Condition: You cannot specify both this and a configuration template name.}
}
\description{
Takes a set of configuration settings and either a configuration template or environment, and determines whether those values are valid.
}
\details{
This action returns a list of messages indicating any errors or warnings associated with the selection of option values.
}
\examples{
# The following operation validates a CloudWatch custom metrics config
# document:
validate_configuration_settings(
  ApplicationName = "my-app",
  EnvironmentName = "my-env",
  OptionSettings = list(
    list(
      Namespace = "aws:elasticbeanstalk:healthreporting:system",
      OptionName = "ConfigDocument",
      Value = "{\\"CloudWatchMetrics\\": {\\"Environment\\": {\\"ApplicationLatencyP99.9\\": null,\\"InstancesSevere\\": 60,\\"ApplicationLatencyP90\\": 60,\\"ApplicationLatencyP99\\": null,\\"ApplicationLatencyP95\\": 60,\\"InstancesUnknown\\": 60,\\"ApplicationLatencyP85\\": 60,\\"InstancesInfo\\": null,\\"ApplicationRequests2xx\\": null,\\"InstancesDegraded\\": null,\\"InstancesWarning\\": 60,\\"ApplicationLatencyP50\\": 60,\\"ApplicationRequestsTotal\\": null,\\"InstancesNoData\\": null,\\"InstancesPending\\": 60,\\"ApplicationLatencyP10\\": null,\\"ApplicationRequests5xx\\": null,\\"ApplicationLatencyP75\\": null,\\"InstancesOk\\": 60,\\"ApplicationRequests3xx\\": null,\\"ApplicationRequests4xx\\": null},\\"Instance\\": {\\"ApplicationLatencyP99.9\\": null,\\"ApplicationLatencyP90\\": 60,\\"ApplicationLatencyP99\\": null,\\"ApplicationLatencyP95\\": null,\\"ApplicationLatencyP85\\": null,\\"CPUUser\\": 60,\\"ApplicationRequests2xx\\": null,\\"CPUIdle\\": null,\\"ApplicationLatencyP50\\": null,\\"ApplicationRequestsTotal\\": 60,\\"RootFilesystemUtil\\": null,\\"LoadAverage1min\\": null,\\"CPUIrq\\": null,\\"CPUNice\\": 60,\\"CPUIowait\\": 60,\\"ApplicationLatencyP10\\": null,\\"LoadAverage5min\\": null,\\"ApplicationRequests5xx\\": null,\\"ApplicationLatencyP75\\": 60,\\"CPUSystem\\": 60,\\"ApplicationRequests3xx\\": 60,\\"ApplicationRequests4xx\\": null,\\"InstanceHealth\\": null,\\"CPUSoftirq\\": 60}},\\"Version\\": 1}"
    )
  )
)

}
