% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{check_dns_availability}
\alias{check_dns_availability}
\title{Checks if the specified CNAME is available}
\usage{
check_dns_availability(
  CNAMEPrefix = "string"
)
}
\arguments{
\item{CNAMEPrefix}{[required] The prefix used when this CNAME is reserved.}
}
\description{
Checks if the specified CNAME is available.
}
\examples{
# The following operation checks the availability of the subdomain
# my-cname:
check_dns_availability(
  CNAMEPrefix = "my-cname"
)

}
