% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{create_platform_version}
\alias{create_platform_version}
\title{Create a new version of your custom platform}
\usage{
create_platform_version(PlatformName, PlatformVersion, PlatformDefinitionBundle,
  EnvironmentName = NULL, OptionSettings = NULL)
}
\arguments{
\item{PlatformName}{The name of your custom platform.}

\item{PlatformVersion}{The number, such as 1.0.2, for the new platform version.}

\item{PlatformDefinitionBundle}{The location of the platform definition archive in Amazon S3.}

\item{EnvironmentName}{The name of the builder environment.}

\item{OptionSettings}{The configuration option settings to apply to the builder environment.}
}
\description{
Create a new version of your custom platform.
}
\examples{

}
