% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elasticbeanstalk_operations.R
\name{update_application_resource_lifecycle}
\alias{update_application_resource_lifecycle}
\title{Modifies lifecycle settings for an application}
\usage{
update_application_resource_lifecycle(ApplicationName, ResourceLifecycleConfig)
}
\arguments{
\item{ApplicationName}{The name of the application.}

\item{ResourceLifecycleConfig}{The lifecycle configuration.}
}
\description{
Modifies lifecycle settings for an application.
}
\examples{

}
